Allow generic type parameter for QuerydslRepositorySupport.
Closes #4300
