+++ spring-data-jpa/src/main/java/org/springframework/data/jpa/repository/support/QuerydslRepositorySupport.java
@@ -15,155 +15,65 @@
  */
 package org.springframework.data.jpa.repository.support;
 
-import jakarta.annotation.PostConstruct;
-import org.jspecify.annotations.Nullable;
-
-import jakarta.persistence.EntityManager;
-
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.ResolvableType;
 import org.springframework.stereotype.Repository;
-import org.springframework.util.Assert;
 
-import com.querydsl.core.dml.DeleteClause;
-import com.querydsl.core.dml.UpdateClause;
-import com.querydsl.core.types.EntityPath;
 import com.querydsl.core.types.dsl.PathBuilder;
-import com.querydsl.core.types.dsl.PathBuilderFactory;
-import com.querydsl.jpa.JPQLQuery;
-import com.querydsl.jpa.impl.JPADeleteClause;
-import com.querydsl.jpa.impl.JPAUpdateClause;
 
 /**
  * Base class for implementing repositories using Querydsl library.
  *
  * @author Oliver Gierke
  * @author Mark Paluch
+ * @author Maxim Shabanov
  */
 @Repository
-public abstract class QuerydslRepositorySupport {
+public abstract non-sealed class QuerydslRepositorySupport<T> extends QuerydslRepositorySupportBase {
 
-	private final PathBuilder<?> builder;
+    /**
+     * Creates a new {@link QuerydslRepositorySupport} instance.
+     * <p>
+     * This constructor automatically detects the domain type {@code T} by resolving the generic parameter of the
+     * concrete subclass.
+     * 
+     * @throws IllegalArgumentException if the domain type {@code T} cannot be resolved from the generic signature of
+     *             the subclass.
+     */
+    public QuerydslRepositorySupport() {
+        super();
+    }
 
-	private @Nullable EntityManager entityManager;
-	private @Nullable Querydsl querydsl;
+    /**
+     * Creates a new {@link QuerydslRepositorySupport} instance for the given domain type.
+     *
+     * @param domainClass must not be {@literal null}.
+     */
+    public QuerydslRepositorySupport(Class<?> domainClass) {
+        super(domainClass);
+    }
 
-	/**
-	 * Creates a new {@link QuerydslRepositorySupport} instance for the given domain type.
-	 *
-	 * @param domainClass must not be {@literal null}.
-	 */
-	public QuerydslRepositorySupport(Class<?> domainClass) {
+    /**
+     * Returns a type-safe {@link PathBuilder} for the configured domain type {@code T}.
+     *
+     * @return the Querdsl {@link PathBuilder}.
+     */
+    @SuppressWarnings("unchecked")
+    @Override
+    protected PathBuilder<T> getBuilder() {
+        return (PathBuilder<T>) super.getBuilder();
+    }
 
-		Assert.notNull(domainClass, "Domain class must not be null");
-		this.builder = new PathBuilderFactory().create(domainClass);
-	}
+    @Override
+    Class<?> resolveDomainClass() {
 
-	/**
-	 * Setter to inject {@link EntityManager}.
-	 *
-	 * @param entityManager must not be {@literal null}.
-	 */
-	@Autowired
-	public void setEntityManager(EntityManager entityManager) {
+        Class<?> domainClass = ResolvableType.forClass(this.getClass()).as(QuerydslRepositorySupport.class)
+                .getGeneric(0).resolve();
 
-		Assert.notNull(entityManager, "EntityManager must not be null");
-		this.querydsl = new Querydsl(entityManager, builder);
-		this.entityManager = entityManager;
-	}
+        if (domainClass == null) {
+            throw new IllegalArgumentException(
+                    String.format("Could not resolve generic parameter for %s", this.getClass()));
+        }
 
-	/**
-	 * Callback to verify configuration. Used by containers.
-	 */
-	@PostConstruct
-	public void validate() {
-		Assert.notNull(entityManager, "EntityManager must not be null");
-		Assert.notNull(querydsl, "Querydsl must not be null");
-	}
-
-	/**
-	 * Returns the {@link EntityManager}.
-	 *
-	 * @return the entityManager
-	 */
-	protected @Nullable EntityManager getEntityManager() {
-		return entityManager;
-	}
-
-	/**
-	 * Returns a fresh {@link JPQLQuery}.
-	 *
-	 * @param paths must not be {@literal null}.
-	 * @return the Querydsl {@link JPQLQuery}.
-	 */
-	protected JPQLQuery<Object> from(EntityPath<?>... paths) {
-		return getRequiredQuerydsl().createQuery(paths);
-	}
-
-	/**
-	 * Returns a {@link JPQLQuery} for the given {@link EntityPath}.
-	 *
-	 * @param path must not be {@literal null}.
-	 * @return
-	 */
-	protected <T> JPQLQuery<T> from(EntityPath<T> path) {
-		return getRequiredQuerydsl().createQuery(path).select(path);
-	}
-
-	/**
-	 * Returns a fresh {@link DeleteClause}.
-	 *
-	 * @param path
-	 * @return the Querydsl {@link DeleteClause}.
-	 */
-	protected DeleteClause<JPADeleteClause> delete(EntityPath<?> path) {
-		return new JPADeleteClause(getRequiredEntityManager(), path);
-	}
-
-	/**
-	 * Returns a fresh {@link UpdateClause}.
-	 *
-	 * @param path
-	 * @return the Querydsl {@link UpdateClause}.
-	 */
-	protected UpdateClause<JPAUpdateClause> update(EntityPath<?> path) {
-		return new JPAUpdateClause(getRequiredEntityManager(), path);
-	}
-
-	/**
-	 * Returns a {@link PathBuilder} for the configured domain type.
-	 *
-	 * @param <T>
-	 * @return the Querdsl {@link PathBuilder}.
-	 */
-	@SuppressWarnings("unchecked")
-	protected <T> PathBuilder<T> getBuilder() {
-		return (PathBuilder<T>) builder;
-	}
-
-	/**
-	 * Returns the underlying Querydsl helper instance.
-	 *
-	 * @return
-	 */
-	protected @Nullable Querydsl getQuerydsl() {
-		return this.querydsl;
-	}
-
-	private Querydsl getRequiredQuerydsl() {
-
-		if (querydsl == null) {
-			throw new IllegalStateException("Querydsl is null");
-		}
-
-		return querydsl;
-	}
-
-	private EntityManager getRequiredEntityManager() {
-
-		if (entityManager == null) {
-			throw new IllegalStateException("EntityManager is null");
-		}
-
-		return entityManager;
-	}
+        return domainClass;
+    }
 }
